Move logic into bfcache-opt-in.php
